Split character builder generation runner
